# ENH: spatial: speed up boolean distances