This patch optimizes the boolean distances. Some of them are no faster, others an order of magnitude. Benchmark script:

```py
from IPython import get_ipython
ipython = get_ipython()

X = np.random.RandomState(42).randn(1000, M) > 0
for metric in ["yule", "dice", "rogerstanimoto", "kulsinski",
               "sokalsneath", "sokalmichener"]:
    print(metric, end=" ")
    ipython.run_line_magic("timeit", "pdist(X, metric)")
```

Results, in milliseconds:

Metric|Old, M=4|Old, M=40|New, M=4|New, M=40
--------|------|------|-----|------
yule|8.66|103.00|8.92|102.00
dice|5.78|70.50|5.66|69.70
rogerstanimoto|8.63|105.00|4.05|7.40
kulsinski|8.44|101.00|5.82|69.30
sokalsneath|8.55|101.00|5.72|69.50
sokalmichener|8.39|106.00|2.12|7.35

(I must admit I'm baffled by the difference between, e.g., sokalsneath and sokalmichener, which have the same loop.)