@larsmans 2016-01-18
For some reason, GCC apparently decides to optimize `(u[i] && v[i])` to a simple AND instruction in some functions, while in others it generates a conditional branch. With all functions updated to use `&` instead of `&&`, here are the new timings:

Metric|Old, M=4|Old, M=40|New, M=4|New, M=40
--------|------|------|-----|------
yule|8.66|103.00|4.71|23.40
dice|5.78|70.50|3.28|12.80
rogerstanimoto|8.63|105.00|4.08|7.43
kulsinski|8.44|101.00|3.77|12.60
sokalsneath|8.55|101.00|3.85|12.90
sokalmichener|8.39|106.00|2.13|7.35