@rgommers 2016-01-18
That looks good. Maybe add the timings as an ``asv`` benchmark?