@larsmans 2016-01-24
@rgommers I need to learn how asv works at some point, but don't wait for me to use it in this PR.