@juliantaylor 2016-01-25
hm gh-5027 is still open, the bool functions shouldn't even be used