@juliantaylor 2016-01-25
I am very surprised the compiler does not make these optimizations itself, I see no reason why it should not but it doesn't. I'll file a bug with gcc.