@larsmans 2016-01-25
I've inserted the additional checks for non-boolean chars that @juliantaylor pointed out; some of the functions get slower, but they're still 3×-5× faster than master for high-d inputs:

Metric|Old, M=4|Old, M=40|New, M=4|New, M=40
--------|------|------|-----|------
yule|8.66|103.00|4.87|36.90
dice|5.78|70.50|3.77|26.60
rogerstanimoto|8.63|105.00|2.65|9.27
kulsinski|8.44|101.00|3.86|27.10
sokalsneath|8.55|101.00|3.87|26.80
sokalmichener|8.39|106.00|2.69|9.61