@juliantaylor 2016-01-26
looks good, very nice improvement, thanks
filed a bug with gcc about its inability to optimize the original code: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=69489