@larsmans 2016-01-26
@rgommers Ready for merge?