@rgommers 2016-01-26
LGTM except for the overlooked ``0.0`` that I commented on.