@rgommers 2016-01-27
Merged, thanks Lars, Julian.